Log update / Logic update
#comment: Can add logs and view them according to the matching tour. Can delete logs from tour. During start, unused images are deleted.

#duration: 4

+++ TourPlanner.DAL/Tour/TourRepository.cs
@@ -53,13 +53,14 @@ namespace TourPlanner.DAL.Tour
             if (dbConnection.State == ConnectionState.Closed)
                 dbConnection.Open();
 
+            if (tourData == null) return;
             dbConnection.Query<TourData>("DELETE FROM tours WHERE TourId = @TourId",
                 new { tourData.TourId },
                 commandType: CommandType.Text);
-
             dbConnection.Query<LogData>("DELETE FROM LOGS WHERE TourId = @TourId",
                 new { tourData.TourId },
                 commandType: CommandType.Text);
+
         }
     }
 }
